The faulty database code
Made several changes:
- Added the third party FMDB, which is an sqlite wrapper. Most sources
say this is the best way to implement sqlite with Swift. Along with
that, added several compiler flags and a library in Build Phases.
There’s also a header file to bridge the FMDB’s Objective-C with the
Swift project.
- Added the database file to the project (mol.db). Most online sources
have a .sqlite file, so I’m not sure if that will be a problem later.
Jason’s database was in .db format.
- From there, made 3 additional classes: DBManager, which is where the
bug is; MoleculeInfo, which creates the type to add to the array
receiving from the database, and Utility, which mainly functions to get
the path to the database.

The problem:
In DBManager, the “database” variable wants to not be nil. I talk a
little more in the comments about what might need to be done here. Some
suggestions I’ve found online would be to make sharedInstance static or
to use an if-statement to guard access. I’ll try to work through those,
but if you know how or if you see any other solutions, let me know.
Thanks!