Update on database: Sets
All the database integration is still in the settings view, so
currently when it runs and settings view is clicked, it accesses the
image of a hardcoded index referring to a molecule. Kat and I reworked
the schema so it’s now organized by sets.

There are two new tables in the database: “sets,” which lists the sets
and the set ids (sid), which are the primary keys; and “darthMol,”
which is table that will join “mol” and “sets” (i.e., the two columns
are sid and mid, matching up which molecules are in which sets).

Note: Only Alkanes (the first 11 molecules in the database) have images
to go with them. The rest of the images from Jason’s original database
that are still in the assets folder are useless with the new sets we’ve
added.