Retrieves sets and displays number; New labels class
Katie: There are two new functions in the DBManager, getSetSize and
getASet. Both of these take an int parameter to specify which set to
retrieve from. getASet is called but not yet utilized in CardView,
assigned to array “setInfo.”
In labels, getSetSize is called and displays the number of molecules in
the set at the top of the card.

Kat: started moving tapping and swiping functionality to new class
“Labels.” Also Kat wrote queries and put them on the drive in Front End
folder, in a doc called “Queries.”