Added new images, changes in setManager and cardViewController to get images from database
Updated with images for carboxylic acids (set 7). Added function in
SetManager to return image id, and changed code in the swipedView
function of cardViewController to get images from database; there is a
bug where the first image displayed is always the first in the
database, but if you go forward and back through the sets it corrects
itself.

ISSUE: It throws a runtime error whenever we try to get anything above
set 4